Updating scope_procedure to define singleton methods rather than set up scopes.
This allows any procs passed in to return something that is not a scope. E.g.

     scope_procedure :published_grouped_by_user, lambda { published.group_by(&:user) }

In the above the result of calling "published_grouped_by_user" should be a hash where the keys are the user record and the values are the post records for that user, but before this change-set the result would be an ActiveRecord::Relation which would ignore the group_by operation.