Handle in and not_in queries using ActiveRecord rather than Arel
Previously modern search logic was using Arel to handle in and not_in queries. When nil is passed as an argument Arel handles it incorrectly and generates incorrect SQL. We are now handling in and not_in with ActiveRecord as it handles nil correctly.

An example of the issue with Arel is Song.language_in(['en', nil])

Active record will realize nil needs to be handled as a special case and generate the SQL
language IN ('en') AND language IS NULL

Arel does not and generates the incorrect SQL
language IN ('en', NULL)