Move unscoped call into searchlogic_active_record_alias, reduce where_values in AR 3 & 4 with :and to get rid of nested data ArelNode structure
Unscoped is only needed when querying with ActiveRecord Relations and not Arel Nodes so we are moving the call into searchlogic_active_record_alias.

The where_values on ActiveRecord Relations in AR 3 and 4 are a nested array of Arel Nodes. This necessitated us calling #flatten. However, in ActiveRecord 5 the ast value takes care of :and for us. By reducing with :and in searchlogic_active_record_alias for AR 3 and 4 we can get rid of the nested data structure and remove the call to flatten that is sometimes unnecessary.