Fix column discovery running too early
In a test environment, or when creating/seeding the database, there are no tables defined in ActiveRecord. This causes an edge case where, if *any* `method_missing` is called (and it gets to MSL), an error will be thrown. This is because accessing `columns_hash` or `column_names` will cause a query to be made that will fail (as it expects the table to already be present in the database).

For getting caught up on the memes, see the following:
- https://github.com/Genius/Rap-Genius/pull/15200
- https://github.com/Genius/heroku-buildpack-ruby/pull/14

Feel free to backport this to Rails 2 - seems pointless since it's On The Way Out (TM) (as it has been for the last decade, but who's counting?!)